@@ -1,13 +1,12 @@
 package com.joniroliveira.bconforto.ui;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.AppCompatEditText;
-import android.support.v7.widget.AppCompatSpinner;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.AppCompatEditText;
+import androidx.appcompat.widget.AppCompatSpinner;
+import androidx.appcompat.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.widget.Adapter;
 import android.widget.ArrayAdapter;
 import android.widget.CheckBox;
 import android.widget.SpinnerAdapter;
@@ -123,13 +122,16 @@ public class MainActivity extends AppCompatActivity {
             }
 
             float hoursPrice;
+            float margin;
 
             Timber.i("Spinner position %s", estimateTypeSpinner.getSelectedItemPosition());
 
             if (estimateTypeSpinner.getSelectedItemPosition() == 0){
                 hoursPrice = Settings.getPriceResale(this);
+                margin = Settings.getMarginResale(this);
             }else {
                 hoursPrice = Settings.getPriceConsumer(this);
+                margin = Settings.getMarginConsumer(this);
             }
 
 
@@ -151,6 +153,7 @@ public class MainActivity extends AppCompatActivity {
                             ((Cloth) clothList.get(clothSpinner.getSelectedItemPosition())).getPrice(),
                             Double.parseDouble(clothMeters.getText().toString()),
                             priceFoam,
+                            margin,
                             discountValue);
 
                     DecimalFormat df = new DecimalFormat("#.###");
