@@ -1,8 +1,6 @@
 package com.joniroliveira.bconforto.ui.adapters;
 
-import android.app.Activity;
-import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -38,6 +36,10 @@ public class ClothRecyclerViewAdapter extends RecyclerView.Adapter<ClothRecycler
     public void updateData(){
         getData();
         notifyDataSetChanged();
+    }
+
+    public Cloth getItem(int position){
+        return clothList.get(position);
     }
 
     @Override
