@@ -10,8 +10,12 @@ import com.joniroliveira.bconforto.data.preferences.Settings;
 
 public class CalculatePrice {
 
-    public static Price calculate(double hours, float hoursPrice, double clothPrice, double quantity, float foamPrice,int discount){
-        double price = (hours * hoursPrice) + (clothPrice * quantity) + foamPrice;
+    public static Price calculate(double hours, float hoursPrice, double clothPrice, double quantity, float foamPrice, float margin,  int discount){
+        float marginCal = 1;
+        if (margin != 0){
+            marginCal = 1 + ( margin /100);
+        }
+        double price = (hours * hoursPrice) + (clothPrice * quantity) + foamPrice * marginCal;
         float aux = (float) (discount/100.0);
         double discountValue = price * aux;
         price = price - discountValue;
